refactor: move form preview logic into a custom hook
